project completed

+++ src/custom/pages/todos/todos.tsx
@@ -1,27 +1,38 @@
-import { useNavigate } from "react-router-dom";
+import { deleteTodo, getAllTodos } from "../../../common/services/todo";
 import { useEffect, useState } from "react";
-import {
-  deleteTodo,
-  getAllTodos,
-  updateTodo,
-} from "../../../common/services/todo";
-import styles from "./todos.module.scss";
+import { useNavigate } from "react-router-dom";
 import { TodoItem } from "../../../common/components/todoItem/todoItem";
+import { TodosPDF } from "./todosPDF";
+import { Loading } from "../../../common/components/layout/ui/loading/loading";
 import { Header } from "../../../common/components/layout/ui/header/header";
+import { Button } from "../../../common/components/layout/ui/button/button";
 import { toast } from "react-toastify";
-import { string } from "yup";
+import { ITodo } from "./helpers";
+import { pdf } from "@react-pdf/renderer";
+import styles from "./todos.module.scss";
 
 export const Todos = () => {
   const [data, setData] = useState<ITodo[]>([]);
+  const [loading, setLoading] = useState(false);
+  const [removeLoading, setRemoveLoading] = useState(false);
   const navigate = useNavigate();
 
   useEffect(() => {
+    if (!window.localStorage.getItem("accessToken")) {
+      navigate("/login");
+      return;
+    }
+
     const allTodos = async () => {
       try {
+        setLoading(true);
         const response = await getAllTodos();
         setData(response.data);
-      } catch (error) {
+      } catch (error: any) {
         console.warn(error);
+        toast.error(error);
+      } finally {
+        setLoading(false);
       }
     };
 
@@ -29,19 +40,60 @@ export const Todos = () => {
   }, []);
 
   const handleToggle = (id: number) => {
-    // const todo = data.find((t) => t.id === id);
+    setData(
+      data.map((el) => {
+        if (el.todo_id == id) {
+          return {
+            ...el,
+            checked: !el.checked,
+          };
+        }
+        return el;
+      })
+    );
   };
 
   const handleDelete = async (id: number) => {
-    const response = await deleteTodo(id.toString());
-    console.log(response);
-    toast.success("başarıyla silindi.");
+    if (removeLoading) {
+      return;
+    }
+
+    try {
+      setRemoveLoading(true);
+      const response = await deleteTodo(id.toString());
+      toast.success(response.message);
+      setData(data.filter((el) => el.todo_id !== id));
+    } catch (error: any) {
+      console.warn(error);
+      toast.error(error);
+    } finally {
+      setRemoveLoading(false);
+    }
   };
 
-  const handlePDF = () => {};
+  const handlePDF = async () => {
+    if (removeLoading) {
+      return;
+    }
 
-  if (!window.localStorage.getItem("accessToken")) {
-    return <div>Giriş yapmalısınız</div>;
+    const blob = await pdf(<TodosPDF data={data} />).toBlob();
+    const linkElement = document.createElement("a");
+    linkElement.href = URL.createObjectURL(blob);
+    linkElement.target = "_blank";
+    linkElement.click();
+    linkElement.remove();
+  };
+
+  const handleCreate = () => {
+    if (removeLoading) {
+      return;
+    }
+
+    navigate("/create");
+  };
+
+  if (loading) {
+    return <Loading />;
   }
 
   return (
@@ -50,36 +102,28 @@ export const Todos = () => {
       <div className={styles.contentWrapper}>
         <div className={styles.header}>
           <h2> TODO APP </h2>
-          <button onClick={handlePDF}> Download PDF </button>
-          <button onClick={() => navigate("/create")}>Create Todo</button>
+          <Button onClick={handlePDF} text="Download PDF" />
+          <Button onClick={handleCreate} text="Create Todo" />
         </div>
         <div className={styles.content}>
           {data.length > 0 ? (
             data.map((todo) => (
               <TodoItem
+                removeLoading={removeLoading}
                 key={todo.todo_id}
                 id={todo.todo_id}
                 text={todo.text}
                 time={todo.time}
                 toggle={handleToggle}
                 remove={handleDelete}
+                checked={todo.checked}
               />
             ))
           ) : (
-            <div className={styles.emptyMessage}>Yapılacaklar listesi boş.</div>
+            <div className={styles.emptyMessage}>The to-do list is empty.</div>
           )}
         </div>
       </div>
     </div>
   );
 };
-
-interface ITodo {
-  todo_id: number;
-  user_id: number;
-  time: string;
-  text: string;
-  checked: boolean;
-  createdAt: Date;
-  updatedAt: Date;
-}
